pipeline/requirements.txt
# cache-bust: 2026-03-31
anthropic
supabase
resend
httpx
python-dotenv
fastapi
uvicorn[standard]
playwright
pypdf
apscheduler
